+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/di/MapperProvider.kt
@@ -164,7 +164,7 @@ internal object MapperProvider {
         get() = BusinessSecurityCodeDisplayDataMapper()
 
     val remediesLinkableMapper: RemediesLinkableMapper
-        get() = RemediesLinkableMapper()
+        get() = RemediesLinkableMapper(Session.getInstance().applicationContext)
 
     val uriToDeepLinkWrapperMapper: UriToDeepLinkWrapperMapper
         get() = UriToDeepLinkWrapperMapper(UriToFromMapper())