+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/explode/ExplodingFragment.kt
@@ -33,8 +33,11 @@ import com.mercadopago.android.px.internal.extensions.isNotNullNorEmpty
 import com.mercadopago.android.px.internal.extensions.visible
 import com.mercadopago.android.px.internal.util.ViewUtils
 import kotlin.math.hypot
+import kotlin.math.min
 
-class ExplodingFragment : Fragment() {
+private const val MAX_LOADING_TIME_MILLIS = 60_000
+
+internal class ExplodingFragment : Fragment() {
 
     private var animator: ObjectAnimator? = null
 
@@ -65,7 +68,7 @@ class ExplodingFragment : Fragment() {
         super.onCreate(savedInstanceState)
         arguments?.let {
             buttonText = it.getCharSequence(ARG_PROGRESS_TEXT)
-            maxLoadingTime = it.getInt(ARG_TIMEOUT)
+            maxLoadingTime = min(it.getInt(ARG_TIMEOUT), MAX_LOADING_TIME_MILLIS)
         } ?: error("Missing explode params")
     }