+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/pay_button/PayButtonFragment.kt
@@ -48,7 +48,7 @@ private const val EXTRA_STATE = "extra_state"
 private const val EXTRA_VISIBILITY = "extra_visibility"
 private const val EXTRA_OBSERVING = "extra_observing"
 
-class PayButtonFragment : BaseFragment(), PayButton.View, SecurityValidationHandler {
+internal class PayButtonFragment : BaseFragment(), PayButton.View, SecurityValidationHandler {
 
     private var buttonStatus = MeliButton.State.NORMAL
     private lateinit var button: MeliButton
@@ -76,7 +76,7 @@ class PayButtonFragment : BaseFragment(), PayButton.View, SecurityValidationHand
         button = view.findViewById(R.id.confirm_button)
         button.setOnClickListener(object : OnSingleClickListener() {
             override fun onSingleClick(v: View?) {
-                viewModel.preparePayment()
+                viewModel.onButtonPressed()
             }
         })
         savedInstanceState?.let {
@@ -93,10 +93,12 @@ class PayButtonFragment : BaseFragment(), PayButton.View, SecurityValidationHand
         updateButtonState()
 
         with(viewModel) {
-            buttonTextLiveData.observe(viewLifecycleOwner,
-                Observer { buttonConfig -> button.text = buttonConfig!!.getButtonText(this@PayButtonFragment.context!!) })
-            cvvRequiredLiveData.observe(viewLifecycleOwner,
-                Observer { params -> params?.let { showSecurityCodeScreen(it) } })
+            buttonTextLiveData.observe(viewLifecycleOwner, Observer { buttonConfig ->
+                button.text = buttonConfig!!.getButtonText(this@PayButtonFragment.context!!)
+            })
+            cvvRequiredLiveData.observe(viewLifecycleOwner, Observer { params ->
+                params?.let { showSecurityCodeScreen(it) }
+            })
             stateUILiveData.observe(viewLifecycleOwner, Observer { state -> state?.let { onStateUIChanged(it) } })
         }
     }