+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/payment_result/remedies/view/RetryPaymentFragment.kt
@@ -13,7 +13,6 @@ import com.mercadopago.android.px.internal.experiments.BadgeVariant
 import com.mercadopago.android.px.internal.extensions.gone
 import com.mercadopago.android.px.internal.extensions.visible
 import com.mercadopago.android.px.internal.features.express.slider.PaymentMethodFragment
-import com.mercadopago.android.px.internal.features.payment_result.remedies.RemediesLinkableMapper
 import com.mercadopago.android.px.internal.features.payment_result.remedies.RemediesPayerCost
 import com.mercadopago.android.px.internal.features.payment_result.remedies.RemediesPaymentMethodMapper
 import com.mercadopago.android.px.internal.view.LinkableTextView
@@ -31,6 +30,7 @@ internal class RetryPaymentFragment : Fragment(), PaymentMethodFragment.Disabled
     private lateinit var paymentMethodDescriptor: PaymentMethodDescriptorView
     private lateinit var paymentMethodTitle: MPTextView
     private lateinit var bottomText: LinkableTextView
+    private val remediesLinkableMapper = MapperProvider.remediesLinkableMapper
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater.inflate(R.layout.px_remedies_retry_payment, container, false)
@@ -55,7 +55,7 @@ internal class RetryPaymentFragment : Fragment(), PaymentMethodFragment.Disabled
             }
 
             it.consumerCredits?.let { consumerCredits ->
-                bottomText.updateModel(RemediesLinkableMapper().mapRemedies(consumerCredits.displayInfo.bottomText))
+                bottomText.updateModel(remediesLinkableMapper.mapRemedies(consumerCredits.displayInfo.bottomText))
             }
         }
         model.cvvModel?.let { cvvRemedy.init(it) } ?: cvvRemedy.gone()