+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/viewmodel/SummaryRowIconDescriptor.kt
@@ -5,6 +5,5 @@ import android.view.View
 internal data class SummaryRowIconDescriptor(
     val drawable: IDetailDrawable,
     val drawableColor: IDetailColor,
-    val listener: View.OnClickListener,
-    val url: String? = null
+    val listener: View.OnClickListener
 )